# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Financial Crime Compliance Platform API
  version: 0.6.0
  description: Portfolio API for authenticated customer-risk assessment, financial transaction monitoring, and explainable alert investigation.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      operationId: healthCheck
      responses:
        "200":
          description: Service is healthy
  /readyz:
    get:
      operationId: readinessCheck
      responses:
        "200":
          description: API is ready and PostgreSQL is reachable
        "503":
          description: PostgreSQL is unavailable
  /v1/customers:
    post:
      operationId: onboardCustomer
      summary: Onboard and risk-assess a customer
      description: Requires the `analyst` or `admin` role. The JWT subject is recorded as the audit actor.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardCustomer"
      responses:
        "201":
          description: Customer created, assessed, and submitted for independent approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400": {description: Malformed JSON}
        "401":
          description: Missing, invalid, or expired bearer token
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "403":
          description: Authenticated role is not allowed to onboard customers
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "422": {description: Customer validation failed}
        "500":
          description: Customer and audit event could not be persisted; the transaction is rolled back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/customers/{customer_id}/approve:
    post:
      operationId: approveCustomer
      summary: Approve a pending customer
      description: Requires the `reviewer` or `admin` role. The original maker cannot approve their own submission.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CustomerID"
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ReviewCustomer"}
      responses:
        "200":
          description: Customer approved and activated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Customer"}
        "400": {description: Malformed JSON}
        "401": {description: Missing, invalid, or expired bearer token}
        "403": {description: Authenticated role cannot review customers}
        "404": {description: Customer not found}
        "409": {description: Maker-checker violation or customer is no longer pending}
        "500": {description: Review could not be persisted; the transaction is rolled back}
  /v1/customers/{customer_id}/reject:
    post:
      operationId: rejectCustomer
      summary: Reject a pending customer
      description: Requires the `reviewer` or `admin` role. The original maker cannot reject their own submission.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CustomerID"
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ReviewCustomer"}
      responses:
        "200":
          description: Customer rejected
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Customer"}
        "400": {description: Malformed JSON}
        "401": {description: Missing, invalid, or expired bearer token}
        "403": {description: Authenticated role cannot review customers}
        "404": {description: Customer not found}
        "409": {description: Maker-checker violation or customer is no longer pending}
        "500": {description: Review could not be persisted; the transaction is rolled back}
  /v1/transactions:
    post:
      operationId: ingestTransaction
      summary: Ingest a financial transaction for an active customer
      description: Requires the `analyst` or `admin` role. Monetary values use integer minor units to avoid floating-point rounding.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/IngestTransaction"}
      responses:
        "201":
          description: Transaction and its audit event were persisted atomically
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TransactionIngestionResult"}
        "400": {description: Malformed JSON}
        "401": {description: Missing, invalid, or expired bearer token}
        "403": {description: Authenticated role cannot ingest transactions}
        "404": {description: Customer not found}
        "409": {description: Customer is not active}
        "422": {description: Transaction validation failed}
        "500": {description: Transaction could not be persisted; the transaction is rolled back}
  /v1/alerts:
    get:
      operationId: listAlerts
      summary: List monitoring alerts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: {type: string, enum: [open, closed]}
      responses:
        "200":
          description: Monitoring alerts ordered newest first
          content:
            application/json:
              schema:
                type: object
                required: [alerts]
                properties:
                  alerts:
                    type: array
                    items: {$ref: "#/components/schemas/Alert"}
        "401": {description: Missing, invalid, or expired bearer token}
        "403": {description: Authenticated role cannot view alerts}
        "422": {description: Invalid status filter}
        "500": {description: Alerts could not be listed}
  /v1/alerts/{alert_id}/close:
    post:
      operationId: closeAlert
      summary: Close an open monitoring alert
      description: Requires the `reviewer` or `admin` role and a closure reason.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: alert_id
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CloseAlert"}
      responses:
        "200":
          description: Alert closed with an atomic audit write
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Alert"}
        "400": {description: Malformed JSON}
        "401": {description: Missing, invalid, or expired bearer token}
        "403": {description: Authenticated role cannot close alerts}
        "404": {description: Alert not found}
        "409": {description: Alert is already closed}
        "422": {description: Closure reason is missing}
        "500": {description: Alert closure could not be persisted; the transaction is rolled back}
components:
  parameters:
    CustomerID:
      in: path
      name: customer_id
      required: true
      schema: {type: string, format: uuid}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 JWT containing `sub`, `role`, `iss`, and `exp`. Supported roles are analyst, reviewer, and admin.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
    OnboardCustomer:
      type: object
      required: [type, legal_name, country_code, risk_factors]
      properties:
        external_ref: {type: string}
        type: {type: string, enum: [individual, company]}
        legal_name: {type: string}
        country_code: {type: string, minLength: 2, maxLength: 2}
        risk_factors:
          $ref: "#/components/schemas/RiskFactors"
    ReviewCustomer:
      type: object
      properties:
        reason: {type: string}
    IngestTransaction:
      type: object
      required: [customer_id, direction, amount_minor, currency, counterparty_country, occurred_at]
      properties:
        external_ref: {type: string}
        customer_id: {type: string, format: uuid}
        direction: {type: string, enum: [inbound, outbound]}
        amount_minor: {type: integer, format: int64, minimum: 1}
        currency: {type: string, minLength: 3, maxLength: 3, example: GBP}
        counterparty_country: {type: string, minLength: 2, maxLength: 2, example: DE}
        occurred_at: {type: string, format: date-time}
    Transaction:
      allOf:
        - $ref: "#/components/schemas/IngestTransaction"
        - type: object
          required: [id, ingested_at, ingested_by]
          properties:
            id: {type: string, format: uuid}
            ingested_at: {type: string, format: date-time}
            ingested_by: {type: string}
    TransactionIngestionResult:
      type: object
      required: [transaction, alerts]
      properties:
        transaction: {$ref: "#/components/schemas/Transaction"}
        alerts:
          type: array
          items: {$ref: "#/components/schemas/Alert"}
    Alert:
      type: object
      required: [id, transaction_id, customer_id, rule_code, rule_version, severity, status, reason_code, description, created_at]
      properties:
        id: {type: string, format: uuid}
        transaction_id: {type: string, format: uuid}
        customer_id: {type: string, format: uuid}
        rule_code: {type: string}
        rule_version: {type: string, example: transaction-monitoring-v1}
        severity: {type: string, enum: [medium, high]}
        status: {type: string, enum: [open, closed]}
        reason_code: {type: string}
        description: {type: string}
        created_at: {type: string, format: date-time}
        closed_at: {type: string, format: date-time}
        closed_by: {type: string}
        closure_reason: {type: string}
    CloseAlert:
      type: object
      required: [reason]
      properties:
        reason: {type: string, minLength: 1}
    RiskFactors:
      type: object
      required: [country_risk, source_of_funds_verified]
      properties:
        country_risk: {type: string, enum: [low, medium, high]}
        pep: {type: boolean}
        sanctions_potential_match: {type: boolean}
        high_risk_industry: {type: boolean}
        complex_ownership: {type: boolean}
        source_of_funds_verified: {type: boolean}
    Customer:
      type: object
      required: [id, type, legal_name, country_code, risk_factors, risk_assessment, status, created_by, created_at]
      properties:
        id: {type: string, format: uuid}
        external_ref: {type: string}
        type: {type: string, enum: [individual, company]}
        legal_name: {type: string}
        country_code: {type: string}
        status: {type: string, enum: [pending_approval, active, rejected]}
        created_by: {type: string}
        reviewed_by: {type: string}
        reviewed_at: {type: string, format: date-time}
        risk_factors: {$ref: "#/components/schemas/RiskFactors"}
        risk_assessment:
          type: object
          properties:
            score: {type: integer}
            rating: {type: string, enum: [low, medium, high]}
            due_diligence: {type: string, enum: [standard, enhanced, blocked_pending_review]}
            reasons:
              type: array
              items:
                type: object
                properties:
                  code: {type: string}
                  points: {type: integer}
                  detail: {type: string}
            assessed_at: {type: string, format: date-time}
            rule_version: {type: string}
        created_at: {type: string, format: date-time}